Implement stack instead of linked-list
At first, I implemented linked-list because I followed the requirements
indicated in the 'one component at a time algorithm of connected-
component labeling I found on Wikipedia.

I ended up implementing a linked-list that behaves like a stack data
structure. And because of that and the size of the expected input is
known, I replaced linked-list with a stack.