Fix non-numeric value input in read_number()
Issue:
In read_number() function, it only return the digits before the non-
numneric value from the input.
    Example:
        Enter a number> 12E45

    The value set in the 'destination' parameter is 12. Which is the
    other values are ignored.

Fix:
To address this issue, I implemented a loop that check each character
in the string input if is a digit or not. If a non-numeric value is
found, read_number() returns an error number and set 'destination'
parameter to NULL.