Fix read_number() function has no error check
At previous version, I decided not implement an error check yet due to
some functions have still issues that need to be addressed first and
read_number() (and read_str() also that it uses under the hood) is one
of them. And now that issues on the function (or on these functions I
should say) are resolved, I implemented an error check for each
read_number() function used.